+++ ChangeLog.md
@@ -1,5 +1,14 @@
 # AgiNG ChangeLog
 
+## Unreleased
+
+**Released: WiP**
+
+- Added the ability to configure the keyboard bindings for the system
+  commands. ([#40](https://github.com/davep/aging/pull/40))
+- Added `--bindings` as a command line switch.
+  ([#40](https://github.com/davep/aging/pull/40))
+
 ## v0.4.0
 
 **Released: 2025-03-26**